Support for JPEG and PNG images
Added support for JPEG and PNG images using the -f --format switch. If either of these formats is specified using the switch, the script will search for and combine images of that format into a PDF, instead of searching for TIFFs. If the switch is not used, the script searches for TIFFs. Renamed the script from "tiffs_to_pdf" to "images_to_pdf" to reflect this change. Updated readme to be format-agnostic. Will update the repository name as well.
